deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
scikit-learn==1.3.1
altair
feature-engine
opencv-python
tensorflow